# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Travel Planner API",
    "description": "여행 기록 관리 서버 API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    }
  ],
  "paths": {
    "/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "loginType": {
                    "example": "any"
                  },
                  "jwt": {
                    "example": "any"
                  },
                  "name": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/schedule": {
      "post": {
        "tags": [
          "Calendar"
        ],
        "description": "",
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "example": "any"
                  },
                  "startDate": {
                    "example": "any"
                  },
                  "endDate": {
                    "example": "any"
                  },
                  "memo": {
                    "example": "any"
                  },
                  "color": {
                    "example": "any"
                  },
                  "shared": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/schedule/{id}": {
      "patch": {
        "tags": [
          "Calendar"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "example": "any"
                  },
                  "startDate": {
                    "example": "any"
                  },
                  "endDate": {
                    "example": "any"
                  },
                  "memo": {
                    "example": "any"
                  },
                  "color": {
                    "example": "any"
                  },
                  "shared": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Schedule"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/todo": {
      "post": {
        "tags": [
          "Calendar"
        ],
        "description": "",
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "example": "any"
                  },
                  "isDone": {
                    "example": "any"
                  },
                  "endDate": {
                    "example": "any"
                  },
                  "memo": {
                    "example": "any"
                  },
                  "color": {
                    "example": "any"
                  },
                  "shared": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/todo/{id}": {
      "patch": {
        "tags": [
          "Calendar"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "example": "any"
                  },
                  "endDate": {
                    "example": "any"
                  },
                  "memo": {
                    "example": "any"
                  },
                  "isDone": {
                    "example": "any"
                  },
                  "color": {
                    "example": "any"
                  },
                  "shared": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Calendar"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/diary": {
      "post": {
        "tags": [
          "Calendar"
        ],
        "description": "",
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date": {
                    "example": "any"
                  },
                  "content": {
                    "example": "any"
                  },
                  "shared": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/diary/{id}": {
      "patch": {
        "tags": [
          "Calendar"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Calendar"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/calendar": {
      "get": {
        "tags": [
          "Calendar"
        ],
        "description": "",
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/trip": {
      "post": {
        "tags": [
          "Trip"
        ],
        "summary": "새 여행 생성",
        "description": "",
        "parameters": [
          {
            "name": "images",
            "in": "formData",
            "items": {
              "type": "file"
            },
            "required": true,
            "description": "여행 사진들",
            "schema": {
              "type": "array"
            }
          },
          {
            "name": "imageOrders",
            "in": "formData",
            "items": {
              "type": "integer"
            },
            "description": "각 사진의 순번",
            "schema": {
              "type": "array"
            }
          },
          {
            "name": "thumbIndex",
            "in": "formData",
            "required": true,
            "description": "썸네일로 사용할 인덱스",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "startDate",
            "in": "formData",
            "format": "date",
            "required": true,
            "description": "여행 시작일 (YYYY-MM-DD)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "in": "formData",
            "format": "date",
            "required": true,
            "description": "여행 종료일 (YYYY-MM-DD)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memo",
            "in": "formData",
            "description": "메모",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sigunguCode",
            "in": "formData",
            "required": true,
            "description": "시군구 코드",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "centerX",
            "in": "formData",
            "format": "float",
            "required": true,
            "description": "지도 중심 경도",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "centerY",
            "in": "formData",
            "format": "float",
            "required": true,
            "description": "지도 중심 위도",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Trip 생성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trip"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Trip"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "imageOrders": {
                    "example": "any"
                  },
                  "thumbIndex": {
                    "example": "any"
                  },
                  "startDate": {
                    "example": "any"
                  },
                  "endDate": {
                    "example": "any"
                  },
                  "memo": {
                    "example": "any"
                  },
                  "sigunguCode": {
                    "example": "any"
                  },
                  "centerX": {
                    "example": "any"
                  },
                  "centerY": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Trip"
        ],
        "description": "",
        "parameters": [
          {
            "name": "lastSync",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/trip/{id}": {
      "patch": {
        "tags": [
          "Trip"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "existing": {
                    "example": "any"
                  },
                  "imageOrders": {
                    "example": "any"
                  },
                  "thumbIndex": {
                    "example": "any"
                  },
                  "startDate": {
                    "example": "any"
                  },
                  "endDate": {
                    "example": "any"
                  },
                  "memo": {
                    "example": "any"
                  },
                  "sigunguCode": {
                    "example": "any"
                  },
                  "centerX": {
                    "example": "any"
                  },
                  "centerY": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Trip"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/tripImage/{filename}": {
      "get": {
        "tags": [
          "Trip"
        ],
        "description": "",
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Trip": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "images": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "array"
                  },
                  "items": {
                    "xml": {
                      "name": "tripimage"
                    },
                    "$ref": "#/components/schemas/TripImage"
                  }
                }
              },
              "startDate": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "date-time"
                  }
                }
              }
            }
          }
        },
        "xml": {
          "name": "Trip"
        }
      },
      "TripImage": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "filename": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "order": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  }
                }
              }
            }
          }
        },
        "xml": {
          "name": "TripImage"
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header",
        "description": "JWT 토큰: Bearer {token}"
      }
    }
  }
}